Fix Driver & Place Dao

+++ blablacar-persistence/src/main/java/cz/fi/muni/pa165/teamred/dao/DriverDaoImpl.java
@@ -4,9 +4,9 @@ import cz.fi.muni.pa165.teamred.entity.Driver;
 import org.springframework.stereotype.Repository;
 
 import javax.persistence.EntityManager;
+import javax.persistence.NoResultException;
 import javax.persistence.PersistenceContext;
 import java.util.List;
-import org.springframework.transaction.annotation.Transactional;
 
 /**
  * @author Erik Horváth
@@ -17,12 +17,18 @@ public class DriverDaoImpl implements DriverDao {
     private EntityManager em;
 
     @Override
-    public void create(Driver driver) {
+    public void create(Driver driver) throws IllegalArgumentException {
+        if(driver == null) {
+            throw new IllegalArgumentException("Driver argument is null.");
+        }
         em.persist(driver);
     }
 
     @Override
-    public void delete(Driver driver) {
+    public void delete(Driver driver) throws IllegalArgumentException {
+        if(driver == null) {
+            throw new IllegalArgumentException("Driver argument is null.");
+        }
         em.remove(em.contains(driver) ? driver : em.merge(driver));
     }
 
@@ -32,7 +38,14 @@ public class DriverDaoImpl implements DriverDao {
     }
 
     @Override
-    public Driver findById(Long id) {
-        return em.find(Driver.class, id);
+    public Driver findById(Long id) throws IllegalArgumentException {
+        if(id == null) {
+            throw new IllegalArgumentException("Id argument is null.");
+        }
+        try {
+            return em.find(Driver.class, id);
+        } catch (NoResultException e) {
+            return null;
+        }
     }
 }
